feat(rust): move credential into main ockam crate
