+++ implementations/rust/ockam/ockam/Cargo.toml
@@ -17,9 +17,14 @@ version = "0.2.0"
 
 [features]
 default = ["std"]
-std = ["ockam_node", "serde/std"]
+std = ["ockam_node", "serde/std", "bbs", "ff", "digest/std", "pairing-plus", "sha2/std"]
+alloc = ["ockam_core/alloc", "serde/alloc"]
+no_std = ["ockam_core/no_std", "serde"]
 
 [dependencies]
+bbs = { version = "0.4", optional = true }
+digest = { version = "0.8", optional = true }
+ff = { version = "0.6", package = "ff-zeroize", optional = true }
 ockam_core = {path = "../ockam_core", version = "0.3.0"}
 ockam_node = {path = "../ockam_node", version = "0.2.0", optional = true}
 ockam_node_attribute = {path = "../ockam_node_attribute", version = "0.1.3"}
@@ -28,9 +33,12 @@ ockam_vault = {path = "../ockam_vault", version = "0.2.0"}
 arrayref = "0.3"
 hex = "0.4"
 hashbrown = { version = "0.9.1", features = ["serde"] }
+pairing-plus = { version = "0.19", optional = true }
 serde_bare = "0.3"
-serde = { version = "1.0", default-features = false, features = ["derive"] }
+serde = { version = "1.0", default-features = false, features = ["derive"], optional = true }
 serde-big-array = "0.3"
+sha2 = { version = "0.8", optional = true }
 
 [dev-dependencies]
 trybuild = {version = "1.0.38", features = ["diff"]}
+serde_json = "1.0"
