+++ implementations/rust/ockam/ockam/src/credential/error.rs
