+++ implementations/rust/ockam/ockam/src/credential/schema.rs
@@ -1,4 +1,4 @@
-use crate::{credential_attribute_schema::CredentialAttributeSchema, serde::*};
+use super::*;
 use ockam_core::lib::*;
 use serde::{Deserialize, Serialize};
