--- implementations/rust/ockam/ockam_credential/Cargo.toml
@@ -1,38 +1,0 @@
-[package]
-name = "ockam_credential"
-version = "0.2.0"
-authors = ["Ockam Developers"]
-edition = "2018"
-license = "Apache-2.0"
-homepage = "https://github.com/ockam-network/ockam"
-repository = "https://github.com/ockam-network/ockam/tree/develop/implementations/rust/ockam/ockam_credential"
-readme = "README.md"
-keywords = ["ockam", "authorization", "anonymous", "zero-knowledge", "credential"]
-categories = ["cryptography", "no-std"]
-description = """
-Attribute based, privacy preserving, anonymous credentials.
-"""
-exclude = [
-    "DEVELOP.md",
-    "LICENSE"
-]
-
-[features]
-default = ["std"]
-std = ["ockam_core/std", "alloc", "bbs", "digest/std", "ff", "pairing-plus", "sha2/std"]
-alloc = ["ockam_core/alloc", "serde/alloc"]
-no_std = ["ockam_core/no_std"]
-
-[dependencies]
-ockam_core = { version = "0.3.0", default-features = false, path = "../ockam_core" }
-bbs = { version = "0.4", optional = true }
-digest = { version = "0.8", optional = true }
-ff = { version = "0.6", package = "ff-zeroize", optional = true }
-heapless = { version = "0.6", optional = true }
-pairing-plus = { version = "0.19", optional = true }
-rand_core = { version = "0.6", default-features = false }
-serde = { version = "1.0", default-features = false, features = ["derive"] }
-sha2 = { version = "0.8", optional = true }
-
-[dev-dependencies]
-serde_json = "1.0"
