--- implementations/rust/ockam/ockam_credential/DEVELOP.md
@@ -1,1 +1,0 @@
-../../DEVELOP.md
