--- implementations/rust/ockam/ockam_credential/LICENSE
@@ -1,1 +1,0 @@
-../../../../LICENSE
